Fix Cython build with Python 3.14
As described in https://bugs.debian.org/1122023, the Cython build fails
with Python 3.14.  This is similar to #23; it has come back because
Python 3.14 changed the default start method to `forkserver`, which has
similar properties to `spawn` in this area.

Rather than continuing to play whack-a-mole by special-casing more start
methods, it's better to take the approach suggested in
https://github.com/akornatskyy/wheezy.template/issues/23#issuecomment-737210953
and move most of `setup.py` into a function.